+++ CozyRead/Views/ReadingList.swift
@@ -14,7 +14,7 @@ private struct Item : View {
     let book: BookCSVData
 
     @State private var expand: Bool = false
-    @State private var rating: Int = 0
+    @State private var rating: Double = 0
 
     var body: some View {
         VStack(alignment: .leading, spacing: 10) {
